Implement the main pipeline
https://github.com/domcermak/neural_style_transfer/issues/1

The pipeline consists 4 apps:
- upload_app
- neural worker
- presention_app
- rabbitmq

Upload app is an app for actual users to upload content image and select style image.
It is prepared to be served publicly through the ngrok proxy. Uploaded images are
queued to be processed by Neural worker.

Neural worker fetches unprocessed images from the queue and processed them with Neural Style Trasfer.
Original images and the generated image by the Neural worker are passed to the processed queue.

Presentation app runs fetches processed images from the proccessed queue and sends them to a locally running browser.

RabbitMQ is a queue managment app.